made changes in cartridge to see if code works

+++ cartridge/include/controller.h
@@ -9,7 +9,7 @@ uint32_t get_controller_sys();
 
 
 
-int8_t get_controller() {
+uint32_t get_controller() {
     return get_controller_sys();
 }
